requirements.txt
bleach==2.1.2
decorator==4.1.2
Django==2.0
django-extensions==1.9.8
entrypoints==0.2.3
html5lib==1.0.1
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.0
Jinja2==2.10
jsonschema==2.6.0
jupyter-core==4.4.0
MarkupSafe==1.0
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
pandocfilters==1.4.2
parso==0.1.0
pexpect==4.3.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
ptyprocess==0.5.2
py2neo==3.1.2
PyGithub==1.35
Pygments==2.2.0
PyJWT==1.5.3
python-dateutil==2.6.1
pytz==2017.3
simplegeneric==0.8.1
six==1.11.0
terminado==0.8.1
testpath==0.3.1
tornado==4.5.2
traitlets==4.3.2
typing==3.6.2
wcwidth==0.1.7
webencodings==0.5.1
